rm setGov.dfip2206f.test.ts, update setGov.test.ts & listPendingDusdSwaps.test.ts
